Перечитал условия и внёс правки
